[REMVIP-9045] fotos subdivision no entran cdn

+++ recovery-webservice/src/main/java/es/pfsgroup/plugin/rem/activo/ActivoManager.java
@@ -1061,7 +1061,7 @@ public class ActivoManager extends BusinessOperationOverrider<ActivoApi> impleme
 
 	@Override
 	@BusinessOperationDefinition("activoManager.getMaxOrdenFotoByIdSubdivision")
-	public Integer getMaxOrdenFotoByIdSubdivision(Long idEntidad, BigDecimal hashSdv) {
+	public Integer getMaxOrdenFotoByIdSubdivision(Long idEntidad, Long hashSdv) {
 		return activoDao.getMaxOrdenFotoByIdSubdivision(idEntidad, hashSdv);
 	}
